% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_pH_spec}
\alias{calc_pH_spec}
\title{Calculate pH values from m-cresol dye absorbance.}
\usage{
calc_pH_spec(
  A730_blank,
  A578_blank,
  A434_blank,
  A730_dye,
  A578_dye,
  A434_dye,
  vol.dye.uL,
  salinity,
  verbose = TRUE,
  calibration = calibration
)
}
\value{
A \link{data.frame} containing calculated pH values with metadata, or a \link{vector} containing just pH values if verbose == FALSE.
}
\description{
This function returns a data.frame containing calculated pH values.
}
\details{
\itemize{
\item A730_blank A \link{vector} or \link{numeric} of samples absorbance at 730nm with no dye.
\item A578_blank A \link{vector} or \link{numeric} of samples absorbance at 578nm with no dye.
\item A434_blank A \link{vector} or \link{numeric} of samples absorbance at 434nm with no dye.
\item A730_dye A \link{vector} or \link{numeric} of samples absorbance at 730nm with m-cresol dye.
\item A578_dye A \link{vector} or \link{numeric} of samples absorbance at 578nm with m-cresol dye.
\item A434_dye A \link{vector} or \link{numeric} of samples absorbance at 434nm with m-cresol dye.
\item vol.dye.L Volumn of dye added in liters (defaults to 0.01).
\item salinity Sample salinity in PSU (defaults to 35).
\item verbose Should only the pH values (False) be returned or metadata as well (True)?
}
}
