% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_plate}
\alias{calc_plate}
\title{Calculate pH values from m-cresol dye absorbance for an entire plate at once.}
\usage{
calc_plate(
  tidyplate,
  verbose = TRUE,
  pairs = data.frame(blanks = c("A", "C", "E", "G"), dyes = c("B", "D", "F", "H")),
  salinty = 35,
  vol.dye.L = 0.01
)
}
\value{
A \link{data.frame} containing calculated pH values with metadata.
}
\description{
This function returns a data.frame containing calculated pH values and metadata.
}
\details{
#' @details text describing parameter inputs in more detail.
\itemize{
\item tidyplate A \link{data.frame} containing formatted absorbance created using \code{\link[=tidyplate]{tidyplate()}}.
\item pairs A \link{data.frame} containing paired 'blank' and 'dye' rows. Defaults too: \link{data.frame(blanks = c("A", "C", "E", "G"), dyes = c("B", "D", "F", "H"))}
\item vol.dye.L Volumn of dye added in liters (defaults to 0.01).
\item salinity Sample salinity in PSU (defaults to 35).
\item verbose Should only the pH values (False) be returned or metadata as well (True)?
}
}
