% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidyplate}
\alias{tidyplate}
\title{Nicely formats plate reader data.}
\usage{
tidyplate(dat)
}
\arguments{
\item{A}{location of raw plate reader data.}
}
\description{
Nicely formats plate reader data.
}
