% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidyplate}
\alias{tidyplate}
\title{Import and format plate reader data from TECAN i-control.}
\usage{
tidyplate(loc)
}
\arguments{
\item{loc}{A \link{character} string containing the file location of raw plate reader data to import.}
}
\value{
A \link{data.frame} containing imported and formatted data.
}
\description{
This function returns a data.frame containing formatted data from a from TECAN i-control csv output.
}
